PublicationPolicy

+++ app/Http/Controllers/PublicationController.php
@@ -6,8 +6,15 @@ use App\Http\Requests\StorePublicationRequest;
 use App\Http\Requests\UpdatePublicationRequest;
 use App\Models\Publication;
 use App\Models\User;
+use Illuminate\Support\Facades\Gate;
+use App\Policies\PublicationPolicy;
 
 class PublicationController extends Controller {
+    /*public function __construct()
+    {
+        $this->authorizeResource(Publication::class, 'publication');
+    }*/
+
     public function index() {
         $publications = Publication::orderBy('created_at', 'desc')->get();
 
@@ -24,11 +31,13 @@ class PublicationController extends Controller {
     }
 
     public function create() {
+        
         $users = User::all();
         return view('publication.form', ['authors' => $users]);
     }
 
     public function add(StorePublicationRequest $request) {
+
         $data = $request->validated();
 
         $newPublication = new Publication($data);
@@ -40,6 +49,9 @@ class PublicationController extends Controller {
     }
 
     public function edit(Publication $publication) {
+
+        $this->authorize('update', $publication);
+
         $users = User::all();
 
         return view(
@@ -52,6 +64,7 @@ class PublicationController extends Controller {
         UpdatePublicationRequest $request,
         Publication              $publication
     ) {
+        $this->authorize('update', $publication);
         $data = $request->validated();
         $publication->fill($data);
         $publication->save();
@@ -62,6 +75,7 @@ class PublicationController extends Controller {
     }
 
     public function destroy(Publication $publication) {
+        $this->authorize('update', $publication);
         $publication->delete();
 
         return redirect()->route(
